// src/app/shared/model/company.model.ts
export class Company {
  key: string;
  name: string;
  address: string;
  imageUrl: string;
  nit: string;
  webAddress: string;
  emailEmergencyContact: string;
  nameEmergencyContact: string;
  phoneEmergencyContact: string;
  legalAddress: string;
  maxBranches: number;
  maxEconomicActivities: number;
  emailRequired: boolean;
}